+++ services/main-website/src/app/page.tsx
@@ -1,12 +1,10 @@
-import { Button } from "@/components/button";
-import {HomepageHero} from "@/components/homepage-hero";
+import { HomepageHero } from "@/components/homepage/hero/hero";
 import styles from "./page.module.scss";
-import "@/styles/global.scss";
 
 export default function Home() {
   return (
     <div className={styles.home}>
-      <HomepageHero/>
+      <HomepageHero />
     </div>
   );
 }
