+++ services/main-website/src/styles/breakpoints.scss
@@ -1,5 +1,5 @@
 $breakpoints: (
-        "phone": 425px,
+        "phone": 440px,
         "tablet": 768px,
         "desktop-sm": 950px,
         "desktop": 1440px,
